xresources: switch urxvt font
Previously used Inconsolata, falling back to Terminus.  But the fallback
failed to work as Inconsolata seems to allegedly provide glyphs for
e.g. the lines in `tree' output, so Terminus is not invoked.

Moved over to xft:Mono which seems to look OK and render UT8-demo.txt
very well/